migtd: move the `exchange_msk` to `bin`
It can only be built for baremetal target and run in baremetal TDX
environment. Move this to bin to avoid lots of conditional compilation
in `migration` module under lib.
