+++ src/migtd/src/migration/session.rs
@@ -40,32 +40,6 @@ impl MigrationInformation {
     }
 }
 
-struct ExchangeInformation {
-    min_ver: u16,
-    max_ver: u16,
-    key: MigrationSessionKey,
-}
-
-impl Default for ExchangeInformation {
-    fn default() -> Self {
-        Self {
-            key: MigrationSessionKey::new(),
-            min_ver: 0,
-            max_ver: 0,
-        }
-    }
-}
-
-impl ExchangeInformation {
-    fn as_bytes(&self) -> &[u8] {
-        unsafe { core::slice::from_raw_parts(self as *const Self as *const u8, size_of::<Self>()) }
-    }
-
-    fn as_bytes_mut(&mut self) -> &mut [u8] {
-        unsafe { core::slice::from_raw_parts_mut(self as *mut Self as *mut u8, size_of::<Self>()) }
-    }
-}
-
 struct VmcallService {
     command: SharedMemory,
     response: SharedMemory,
@@ -307,100 +281,28 @@ pub fn shutdown() -> Result<()> {
     vmcall_service.call()
 }
 
-#[cfg(feature = "main")]
-pub fn exchange_msk(info: &MigrationInformation) -> Result<()> {
-    let mut msk = MigrationSessionKey::new();
+/// Used to read a TDX Module global-scope metadata field.
+///
+/// Details can be found in TDX Module v1.5 ABI spec section 'TDG.SYS.RD Leaf'.
+pub fn tdcall_sys_rd(field_identifier: u64) -> core::result::Result<(u64, u64), TdCallError> {
+    const TDVMCALL_SYS_RD: u64 = 0x0000b;
 
-    read_msk(&info.mig_info, &mut msk)?;
-
-    let transport;
-    #[cfg(feature = "virtio-serial")]
-    {
-        use virtio_serial::VirtioSerialPort;
-        const VIRTIO_SERIAL_PORT_ID: u32 = 1;
-
-        let port = VirtioSerialPort::new(VIRTIO_SERIAL_PORT_ID);
-        port.open()?;
-        transport = port;
-    };
-
-    #[cfg(not(feature = "virtio-serial"))]
-    {
-        use vsock::{stream::VsockStream, VsockAddr};
-        // Establish the vsock connection with host
-        let mut vsock = VsockStream::new()?;
-        vsock.connect(&VsockAddr::new(
-            info.mig_socket_info.mig_td_cid as u32,
-            info.mig_socket_info.mig_channel_port,
-        ))?;
-
-        transport = vsock;
-    };
-
-    let mut remote_information = ExchangeInformation::default();
-    let mut exchange_information = ExchangeInformation {
-        key: msk,
+    let mut args = TdcallArgs {
+        rax: TDVMCALL_SYS_RD,
+        rdx: field_identifier,
         ..Default::default()
     };
 
-    // Establish TLS layer connection and negotiate the MSK
-    if info.is_src() {
-        let min_export_version = tdcall_sys_rd(GSM_FIELD_MIN_EXPORT_VERSION)?.1;
-        let max_export_version = tdcall_sys_rd(GSM_FIELD_MAX_EXPORT_VERSION)?.1;
-        if min_export_version > u16::MAX as u64 || max_export_version > u16::MAX as u64 {
-            return Err(MigrationResult::InvalidParameter);
-        }
-        exchange_information.min_ver = min_export_version as u16;
-        exchange_information.max_ver = max_export_version as u16;
+    let ret = td_call(&mut args);
 
-        // TLS client
-        let mut ratls_client =
-            crate::ratls::client(transport).map_err(|_| MigrationResult::SecureSessionError)?;
-
-        // MigTD-S send Migration Session Forward key to peer
-        ratls_client.write(exchange_information.as_bytes())?;
-        let size = ratls_client.read(remote_information.as_bytes_mut())?;
-        if size < size_of::<ExchangeInformation>() {
-            return Err(MigrationResult::NetworkError);
-        }
-    } else {
-        let min_import_version = tdcall_sys_rd(GSM_FIELD_MIN_IMPORT_VERSION)?.1;
-        let max_import_version = tdcall_sys_rd(GSM_FIELD_MAX_IMPORT_VERSION)?.1;
-        if min_import_version > u16::MAX as u64 || max_import_version > u16::MAX as u64 {
-            return Err(MigrationResult::InvalidParameter);
-        }
-        exchange_information.min_ver = min_import_version as u16;
-        exchange_information.max_ver = max_import_version as u16;
-
-        // TLS server
-        let mut ratls_server =
-            crate::ratls::server(transport).map_err(|_| MigrationResult::SecureSessionError)?;
-
-        ratls_server.write(exchange_information.as_bytes())?;
-        let size = ratls_server.read(remote_information.as_bytes_mut())?;
-        if size < size_of::<ExchangeInformation>() {
-            return Err(MigrationResult::NetworkError);
-        }
+    if ret != TDCALL_STATUS_SUCCESS {
+        return Err(ret.into());
     }
 
-    let mig_ver = cal_mig_version(
-        info.is_src(),
-        exchange_information.min_ver,
-        exchange_information.max_ver,
-        remote_information.min_ver,
-        remote_information.max_ver,
-    )?;
-    set_mig_version(info, mig_ver)?;
-
-    write_msk(&info.mig_info, &remote_information.key)?;
-    log::info!("Set MSK and report status\n");
-    exchange_information.key.clear();
-    remote_information.key.clear();
-
-    Ok(())
+    Ok((args.rdx, args.r8))
 }
 
-fn cal_mig_version(
+pub fn cal_mig_version(
     is_src: bool,
     local_min_ver: u16,
     local_max_ver: u16,
@@ -424,7 +326,7 @@ fn cal_mig_version(
     Ok(core::cmp::min(max_export, max_import))
 }
 
-fn set_mig_version(info: &MigrationInformation, mig_ver: u16) -> Result<()> {
+pub fn set_mig_version(info: &MigrationInformation, mig_ver: u16) -> Result<()> {
     tdcall_servtd_wr(
         info.mig_info.binding_handle,
         TDCS_FIELD_MIG_VERSION,
@@ -434,7 +336,7 @@ fn set_mig_version(info: &MigrationInformation, mig_ver: u16) -> Result<()> {
     Ok(())
 }
 
-fn read_msk(mig_info: &MigtdMigrationInformation, msk: &mut MigrationSessionKey) -> Result<()> {
+pub fn read_msk(mig_info: &MigtdMigrationInformation, msk: &mut MigrationSessionKey) -> Result<()> {
     for idx in 0..msk.fields.len() {
         let ret = tdx::tdcall_servtd_rd(
             mig_info.binding_handle,
@@ -447,7 +349,10 @@ fn read_msk(mig_info: &MigtdMigrationInformation, msk: &mut MigrationSessionKey)
     Ok(())
 }
 
-fn write_msk(mig_info: &MigtdMigrationInformation, msk_peer: &MigrationSessionKey) -> Result<()> {
+pub fn write_msk(
+    mig_info: &MigtdMigrationInformation,
+    msk_peer: &MigrationSessionKey,
+) -> Result<()> {
     for idx in 0..msk_peer.fields.len() {
         tdx::tdcall_servtd_wr(
             mig_info.binding_handle,
@@ -459,27 +364,6 @@ fn write_msk(mig_info: &MigtdMigrationInformation, msk_peer: &MigrationSessionKe
     }
 
     Ok(())
-}
-
-/// Used to read a TDX Module global-scope metadata field.
-///
-/// Details can be found in TDX Module v1.5 ABI spec section 'TDG.SYS.RD Leaf'.
-fn tdcall_sys_rd(field_identifier: u64) -> core::result::Result<(u64, u64), TdCallError> {
-    const TDVMCALL_SYS_RD: u64 = 0x0000b;
-
-    let mut args = TdcallArgs {
-        rax: TDVMCALL_SYS_RD,
-        rdx: field_identifier,
-        ..Default::default()
-    };
-
-    let ret = td_call(&mut args);
-
-    if ret != TDCALL_STATUS_SUCCESS {
-        return Err(ret.into());
-    }
-
-    Ok((args.rdx, args.r8))
 }
 
 #[cfg(test)]
